tarea/requirements.txt
fastapi
uvicorn
redis
grpcio
grpcio-tools
pandas
requests


